requirements.txt
botocore==1.12.86
boto3==1.9.4
click==6.7
falcon==1.4.1
fastparquet==0.1.5
gunicorn==19.9.0
matplotlib==3.0.2
numpy==1.15.1
pandas==0.23.4
PyAthena==1.3.0
python-dotenv==0.9.1
python-snappy==0.5.3
raven==6.9.0
s3fs==0.1.5
seaborn==0.9.0
scikit-learn==0.19.2
toml==0.10.0
yattag==1.10.1

git+https://${GITHUB_TOKEN}@github.com/hellofresh/data-science-util-python@1.2.8#egg=hfds
